refactor: multi module 구조로 변경 chat
