+++ chat/src/main/java/focandlol/chat/controller/ChatController.java
@@ -1,28 +1,24 @@
 package focandlol.chat.controller;
 
-import com.campfiredev.growtogether.chat.ChatManager;
-import com.campfiredev.growtogether.chat.DeleteRedis;
-import com.campfiredev.growtogether.chat.RedisChatMockDataGenerator;
-import com.campfiredev.growtogether.chat.dto.ChatMessageDto;
-import com.campfiredev.growtogether.chat.dto.SliceMessageDto;
-import com.campfiredev.growtogether.chat.service.ChatService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import focandlol.chat.ChatManager;
+import focandlol.chat.DeleteRedis;
+import focandlol.chat.RedisChatMockDataGenerator;
+import focandlol.domain.dto.study.chat.ChatMessageDto;
+import focandlol.domain.dto.study.chat.SliceMessageDto;
+import focandlol.chat.service.ChatService;
 import java.time.LocalDateTime;
 import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.ResponseEntity;
-import org.springframework.messaging.MessageHeaders;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
-import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
-import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.messaging.simp.SimpMessageHeaderAccessor;
 import org.springframework.messaging.simp.SimpMessageType;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
