+++ domain/src/main/java/focandlol/domain/dto/study/chat/SliceMessageDto.java
@@ -1,6 +1,5 @@
-package com.campfiredev.growtogether.chat.dto;
+package focandlol.domain.dto.study.chat;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
