+++ domain/src/main/java/focandlol/domain/repository/chat/ChatRepository.java
@@ -1,6 +1,6 @@
-package com.campfiredev.growtogether.chat.repository;
+package focandlol.domain.repository.chat;
 
-import com.campfiredev.growtogether.chat.entity.ChatEntity;
+import focandlol.domain.entity.chat.ChatEntity;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
